% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlvr.R
\name{format_marginals_fixed}
\alias{format_marginals_fixed}
\title{Format Marginals for Fixed Effects}
\usage{
format_marginals_fixed(B_sims, X, Y)
}
\arguments{
\item{B_sims}{A 3D array containing the posterior simulations of the fixed effects.}

\item{X}{Matrix of covariates (predictors) used in the model.}

\item{Y}{Matrix of response variables.}
}
\value{
A list of data frames for each response variable.
}
\description{
This function formats the posterior simulations of the fixed effects
into a list of data frames, one for each response variable.
}
\author{
Joaquín Martínez-Minaya \email{jmarmin@eio.upv.es}
}
