% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlvr.R
\name{format_marginals_random}
\alias{format_marginals_random}
\title{Format Marginals for Random Effects (Hyperparameters)}
\usage{
format_marginals_random(Sigma_sims)
}
\arguments{
\item{Sigma_sims}{A 3D array containing the posterior simulations of the covariance matrix.}
}
\value{
A data frame with rows as simulations and columns as elements of the covariance matrix.
}
\description{
This function formats the posterior simulations of the hyperparameters (covariance matrix)
into a single data frame, including only the diagonal and upper-triangle elements.
}
\author{
Joaquín Martínez-Minaya \email{jmarmin@eio.upv.es}
}
