% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlvr.R
\name{mlvr}
\alias{mlvr}
\title{Main function to fit the Bayesian Multivariate Linear Model}
\usage{
mlvr(formula, data, priors = list(), n_sims = 1000)
}
\arguments{
\item{formula}{Formula to specify the model.}

\item{data}{A data frame containing the covariates and response variables.}

\item{priors}{A list specifying the prior parameters. Default is `list()`.}

\item{n_sims}{Number of posterior simulations. Default is 1000.}
}
\value{
A list containing summaries, simulations of fixed effects and hyperparameters, and the fitted model.
}
\description{
This function fits a Bayesian multivariate linear regression model and returns a detailed output
including summaries and simulations of fixed effects and hyperparameters.
}
\details{
## Priors:
- `A`: Precision matrix for `B`. Default is an identity matrix of size `k`.
- `V0`: Scale matrix for the covariance matrix `Sigma`. Default is an identity matrix of size `m`.
- `nu0`: Degrees of freedom for the inverse-Wishart distribution. Default is `m + 2`.
- `B0`: Prior mean for the coefficients `B`. Default is a zero matrix.

## Posteriors:
- The posterior of `B` follows a multivariate normal distribution.
- The posterior of `Sigma` follows an inverse-Wishart distribution.
}
\examples{
\dontrun{
data <- data.frame(X1 = rnorm(100), X2 = rnorm(100), Y1 = rnorm(100), Y2 = rnorm(100))
formula <- as.matrix(data[, c("Y1", "Y2")]) ~ X1 + X2
fit <- mlvr(formula, data)
summary(fit)
plot(fit)
}
}
\author{
Joaquín Martínez-Minaya \email{jmarmin@eio.upv.es}
}
