% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlvr.R
\name{plot.mlvr}
\alias{plot.mlvr}
\title{Plot Posterior Distributions for a Bayesian Multivariate Linear Model}
\usage{
\method{plot}{mlvr}(x, ...)
}
\arguments{
\item{x}{An object of class `mlvr` containing posterior simulations of the model.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A list of ggplot objects: one for the fixed effects and one for the hyperparameters.
}
\description{
This function generates density plots for the posterior distributions of the fixed effects
and hyperparameters of a fitted Bayesian multivariate linear model.
}
