% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlvr.R
\name{predict.mlvr}
\alias{predict.mlvr}
\title{Predictive Distribution for Bayesian Multivariate Linear Model (Analytical and Simulated)}
\usage{
\method{predict}{mlvr}(object, newdata, n_sims = 1000)
}
\arguments{
\item{object}{An object of class `mlvr` containing posterior simulations of the model.}

\item{newdata}{A data frame containing new observations for which predictions are to be made.}

\item{n_sims}{Number of predictive simulations (optional, default is 1000).}
}
\value{
A list containing summaries, analytical and simulated predictive distributions, and marginals.
}
\description{
This function generates both analytical and simulated predictive distributions for new data
based on the fitted Bayesian multivariate linear model.
}
\author{
Joaquín Martínez-Minaya \email{jmarmin@eio.upv.es}
}
