% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlvr.R
\name{summarize_fixed_effects}
\alias{summarize_fixed_effects}
\title{Summarize Fixed Effects for Bayesian Multivariate Linear Model}
\usage{
summarize_fixed_effects(B_sims, X, Y)
}
\arguments{
\item{B_sims}{A 3D array containing the posterior simulations of the fixed effects.}

\item{X}{Matrix of covariates (predictors) used in the model.}

\item{Y}{Matrix of response variables.}
}
\value{
A summary of the fixed effects (mean, sd, quantiles) for each predictor and response variable.
}
\description{
This function summarizes the fixed effects (marginal B | Y, X) from the posterior simulations.
}
\author{
Joaquín Martínez-Minaya \email{jmarmin@eio.upv.es}
}
