% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlvr.R
\name{summarize_hyperparameters}
\alias{summarize_hyperparameters}
\title{Summarize Hyperparameters for Bayesian Multivariate Linear Model}
\usage{
summarize_hyperparameters(Sigma_sims)
}
\arguments{
\item{Sigma_sims}{A 3D array containing the posterior simulations of the covariance matrix.}
}
\value{
A summary of the hyperparameters (mean, sd, quantiles) for each element of the covariance matrix.
}
\description{
This function summarizes the hyperparameters (covariance matrix elements)
from the posterior simulations.
}
\author{
Joaquín Martínez-Minaya \email{jmarmin@eio.upv.es}
}
