% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlvr.R
\name{summary.mlvr}
\alias{summary.mlvr}
\title{Summary of a Bayesian Multivariate Linear Model}
\usage{
\method{summary}{mlvr}(object, ...)
}
\arguments{
\item{object}{An object of class mlvr, containing posterior simulations of the model.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A printed summary of the model fit.
}
\description{
This function provides a summary of the fitted Bayesian multivariate linear model.
It includes summaries of both the marginal and conditional posterior distributions of B,
as well as summaries of the hyperparameters (covariance matrix).
}
\examples{
\dontrun{
summary(fit)  # Summarize the fitted model
}
}
\author{
Joaquín Martínez-Minaya \email{jmarmin@eio.upv.es}
}
